move into subfolder utils to test Vite build
