+++ modules/electron/src/logging.ts
@@ -5,9 +5,9 @@ import {ConsoleAppender} from '@mburchard/bit-log/dist/appender/ConsoleAppender.
 import {FileAppender} from '@mburchard/bit-log/dist/appender/FileAppender.js';
 import {LogLevel} from '@mburchard/bit-log/dist/definitions.js';
 import {app} from 'electron';
-import {getLogPath} from './electron-utils.js';
-import {fileExists, mkDir} from './file-utils.js';
 import {registerFrontendListener} from './ipc.js';
+import {getLogPath} from './utils/electron-utils.js';
+import {fileExists, mkDir} from './utils/file-utils.js';
 import 'source-map-support/register.js';
 
 // temporarily change the log level of the ROOT logger to DEBUG
