+++ modules/electron/src/utils/electron-utils.ts
