+++ modules/electron/src/utils/file-utils.ts
